Remove up and down keys
My reasoning for this is because when using in production, let's say you have a site scrolls when you go up and down with arrow keys, normal users would like to move the site up and down, not the carousel. Removing them entirely is an easy solution, you can get fancy and add another boolean for up and down keys specifically.