[chore] fix inconsistent component state updates
React component state updates using setState may
asynchronously update this.props and this.state, thus it is
not safe to use either of the two when calculating the new
state passed to setState, and instead the callback-based
variant should be used instead.

There is one more instance remaining, where the a state property is set to the
value of itself, which I thought should be discussed before fixing.