fix: use closure to unbind image load listener
On initial mount, a "load" event listener may be
applied to the image returned from
Carousel#getInitialImage. Carousel#getInitialImage
relies on props.selectedItem to find the initial
image in the DOM. It is possible the Carousel
consumer has changed props.selectedItem by the
time unbindEvents is called, so Carousel#getInitialImage
may not return the same image and so the initial load
event listener is not removed. To prevent such, access
the initial image from mount in a closure to remove
the listener.